# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/achievements": {
      "post": {
        "description": "Records a Quran recitation achievement for a student. `percentage_score` is computed on the frontend (from raw counts and the halaqa evaluation settings) and stored as-is. Pass `approve: true` to approve in the same request — caller must have halaqa scope (403 if not, never silently skipped). If the student is absent the request is rejected (400).",
        "operationId": "AchievementsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAchievementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AchievementDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error, invalid verse range, or student absent."
          },
          "403": {
            "description": "No halaqa scope."
          },
          "404": {
            "description": "Not found or out of scope."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Record an achievement",
        "tags": [
          "Achievements"
        ]
      },
      "get": {
        "description": "Returns a paginated list of achievements visible to the caller.\n- **principal / vice_principal**: all achievements in the school.\n- **supervisor**: achievements for halaqat they supervise.\n- **teacher**: achievements for halaqat they are assigned to.\n- **parent**: achievements for their linked students only (read-only, full detail).",
        "operationId": "AchievementsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Defaults to 1.",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Defaults to 20, capped at 100.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "student_id",
            "required": false,
            "in": "query",
            "description": "Filter by student.",
            "schema": {
              "example": 42,
              "type": "number"
            }
          },
          {
            "name": "halaqa_id",
            "required": false,
            "in": "query",
            "description": "Filter by halaqa.",
            "schema": {
              "example": 3,
              "type": "number"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "Filter by exact date.",
            "schema": {
              "format": "date",
              "example": "2026-05-11",
              "type": "string"
            }
          },
          {
            "name": "track_type",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Hifz",
              "type": "string",
              "enum": [
                "Hifz",
                "Near",
                "Far"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "example": "approved",
              "type": "string",
              "enum": [
                "approved",
                "unapproved"
              ]
            }
          },
          {
            "name": "recorded_by",
            "required": false,
            "in": "query",
            "description": "Filter by recorder user ID. Not available for parent role (400).",
            "schema": {
              "example": 7,
              "type": "number"
            }
          },
          {
            "name": "approved_by",
            "required": false,
            "in": "query",
            "description": "Filter by approver user ID. Not available for parent role (400).",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AchievementListData"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List achievements (role-scoped)",
        "tags": [
          "Achievements"
        ]
      }
    },
    "/achievements/{id}": {
      "get": {
        "description": "Returns a single achievement with its recitation positions and itemized errors. Out-of-scope or cross-school access returns 404 (never 403). All roles that can read the achievement (including **parent**) receive the full payload.",
        "operationId": "AchievementsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Achievement ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AchievementDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found or out of scope."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get achievement detail",
        "tags": [
          "Achievements"
        ]
      },
      "patch": {
        "description": "Updates mutable fields on an **unapproved** achievement. Returns 400 if the achievement is approved (unapprove it first). `percentage_score` is updated only when the client sends it (recomputed on the frontend). If surah/verse fields change, the range is re-validated.",
        "operationId": "AchievementsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Achievement ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAchievementDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AchievementDto"
                }
              }
            }
          },
          "400": {
            "description": "Achievement is approved, or invalid verse range."
          },
          "404": {
            "description": "Not found or no halaqa scope."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update an achievement",
        "tags": [
          "Achievements"
        ]
      },
      "delete": {
        "description": "Soft-deletes an achievement.\n- **Unapproved**: any role with halaqa scope can delete.\n- **Approved**: principal only (VP and teachers are rejected with 403).\nIf the deleted achievement was approved, plan item reconciliation runs automatically.",
        "operationId": "AchievementsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Achievement ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "403": {
            "description": "Approved achievement — principal only."
          },
          "404": {
            "description": "Not found or no halaqa scope."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete (soft-delete) an achievement",
        "tags": [
          "Achievements"
        ]
      }
    },
    "/achievements/{id}/approve": {
      "post": {
        "description": "Marks the achievement as approved. Caller must be: principal, VP, supervisor in scope, or any teacher assigned to the halaqa. Returns 400 if already approved. Triggers plan item reconciliation.",
        "operationId": "AchievementsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Achievement ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AchievementDto"
                }
              }
            }
          },
          "400": {
            "description": "Already approved."
          },
          "403": {
            "description": "No halaqa scope."
          },
          "404": {
            "description": "Not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Approve an achievement",
        "tags": [
          "Achievements"
        ]
      }
    },
    "/achievements/{id}/unapprove": {
      "post": {
        "description": "Reverts an approved achievement back to unapproved. Caller must be principal, VP, supervisor in scope, or any teacher assigned to the halaqa. `approved_by` and `approved_at` are preserved for audit purposes (status only is flipped). Returns 400 if not currently approved. Triggers plan item reconciliation.",
        "operationId": "AchievementsController_unapprove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Achievement ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AchievementDto"
                }
              }
            }
          },
          "400": {
            "description": "Not currently approved."
          },
          "403": {
            "description": "No halaqa scope."
          },
          "404": {
            "description": "Not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Unapprove an achievement",
        "tags": [
          "Achievements"
        ]
      }
    },
...
  },
  "info": {
    "title": "Halaqa Backend API",
    "description": "Authentication, user management, and school operations for the Halaqa system. All JSON responses are wrapped in a uniform envelope: `{ \"code\": <http>, \"data\": ... }` on success, `{ \"code\": <http>, \"message\": \"...\" }` on errors. Refresh tokens travel as HttpOnly cookies on `/auth/*`; access tokens use Bearer auth.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
...
      "CreateAchievementDto": {
        "type": "object",
        "properties": {
          "student_id": {
            "type": "number",
            "example": 42,
            "description": "Student ID. Must belong to the caller's school."
          },
          "halaqa_id": {
            "type": "number",
            "example": 3,
            "description": "Halaqa ID. Caller must have scope on this halaqa."
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2026-05-11"
          },
          "track_type": {
            "type": "string",
            "enum": [
              "Hifz",
              "Near",
              "Far"
            ],
            "example": "Hifz"
          },
          "completion_method": {
            "type": "string",
            "enum": [
              "quick",
              "mushaf"
            ],
            "default": "quick",
            "description": "How the achievement was entered. Defaults to `quick`."
          },
          "recitation_method": {
            "type": "string",
            "enum": [
              "full",
              "test"
            ],
            "default": "full",
            "description": "How it was recited. `full` auto-creates one position spanning the whole range. `test` requires `test_positions` (each within the achievement range). Defaults to `full`."
          },
          "test_positions": {
            "description": "Required (>=1) when `recitation_method = test`. Rejected for `full`. Each position must fall within the achievement range and carries its own error counts; the achievement-level counts are their sum.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AchievementTestPositionDto"
            }
          },
          "start_surah": {
            "type": "number",
            "example": 1,
            "minimum": 1,
            "maximum": 114,
            "description": "Starting surah (1–114)."
          },
          "start_verse": {
            "type": "number",
            "example": 1,
            "minimum": 1
          },
          "end_surah": {
            "type": "number",
            "example": 1,
            "minimum": 1,
            "maximum": 114,
            "description": "Ending surah. Must be >= start_surah."
          },
          "end_verse": {
            "type": "number",
            "example": 7,
            "minimum": 1
          },
          "errors": {
            "description": "Errors for `recitation_method = full` only — they attach to the single auto-created position. Rejected for `test` (send errors inside each `test_positions` entry). Each error must fall within the achievement range. The per-type counts are derived from this list.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PositionErrorDto"
            }
          },
          "percentage_score": {
            "type": "number",
            "example": 92.5,
            "minimum": 0,
            "maximum": 100,
            "description": "Computed on the frontend from raw counts and the halaqa evaluation settings. Stored as-is."
          },
          "teacher_notes": {
            "type": "object",
            "nullable": true,
            "example": "Strong on the last 3 verses.",
            "maxLength": 1000
          },
          "approve": {
            "type": "boolean",
            "default": false,
            "description": "Approve in the same request. Caller must have halaqa scope. Lacks scope → 403 (never silently downgraded to unapproved)."
          }
        },
        "required": [
          "student_id",
          "halaqa_id",
          "date",
          "track_type",
          "start_surah",
          "start_verse",
          "end_surah",
          "end_verse",
          "percentage_score"
        ]
      },
...
      "AchievementDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "student_id": {
            "type": "number",
            "example": 42
          },
          "student_name": {
            "type": "object",
            "nullable": true,
            "example": "Yusuf Ahmad",
            "description": "Student's name (denormalized for display)."
          },
          "halaqa_id": {
            "type": "number",
            "example": 3
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2026-05-11"
          },
          "track_type": {
            "type": "string",
            "enum": [
              "Hifz",
              "Near",
              "Far"
            ],
            "example": "Hifz"
          },
          "completion_method": {
            "type": "string",
            "enum": [
              "quick",
              "mushaf"
            ],
            "example": "quick"
          },
          "recitation_method": {
            "type": "string",
            "enum": [
              "full",
              "test"
            ],
            "example": "full"
          },
          "recitation_positions": {
            "description": "Recited/tested positions. One full-range row for `full`; one per tested position for `test`.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecitationPositionDto"
            }
          },
          "start_surah": {
            "type": "number",
            "example": 1
          },
          "start_verse": {
            "type": "number",
            "example": 1
          },
          "end_surah": {
            "type": "number",
            "example": 2
          },
          "end_verse": {
            "type": "number",
            "example": 7
          },
          "mistakes_count": {
            "type": "object",
            "example": 2,
            "nullable": true
          },
          "warnings_count": {
            "type": "object",
            "example": 1,
            "nullable": true
          },
          "tajweed_errors_count": {
            "type": "object",
            "example": 0,
            "nullable": true
          },
          "harakat_errors_count": {
            "type": "object",
            "example": 3,
            "nullable": true,
            "description": "Total harakat (حركات) errors across all positions."
          },
          "percentage_score": {
            "type": "string",
            "example": "95.50",
            "description": "Computed percentage score (2 decimal places)."
          },
          "status": {
            "type": "string",
            "enum": [
              "approved",
              "unapproved"
            ],
            "example": "unapproved"
          },
          "recorded_by_name": {
            "type": "object",
            "nullable": true,
            "example": "Ahmad Ali"
          },
          "approved_by_name": {
            "type": "object",
            "nullable": true,
            "example": "Ahmad Ali"
          },
          "approved_at": {
            "type": "object",
            "nullable": true,
            "example": "2026-05-11T12:00:00.000Z"
          },
          "teacher_notes": {
            "type": "object",
            "nullable": true,
            "example": "Strong on the last 3 verses."
          },
          "created_at": {
            "type": "string",
            "example": "2026-05-11T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "student_id",
          "halaqa_id",
          "date",
          "track_type",
          "completion_method",
          "recitation_method",
          "recitation_positions",
          "start_surah",
          "start_verse",
          "end_surah",
          "end_verse",
          "mistakes_count",
          "warnings_count",
          "tajweed_errors_count",
          "harakat_errors_count",
          "percentage_score",
          "status",
          "recorded_by_name",
          "approved_by_name",
          "approved_at",
          "teacher_notes",
          "created_at"
        ]
      },
      "AchievementListData": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AchievementDto"
            }
          },
          "total": {
            "type": "number",
            "example": 42
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ]
      },
...
    }
  }
}